+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.33.2...HEAD)
 
 ### Added
+
+- `tnaLogo` macro added in `nationalarchives/templates/partials/logo.njk`
+
 ### Changed
 
 - Plain buttons are now underlined